Sph: merge static and modulo grid *.cl, fix rigid body spinning.
Fix: Although the mapRigidFluidContacts() kernel compiles, directly
copying FluidRigidContacts(644 bytes) likely exceeds the amount
of available memory. The result is that the copy in the private memory
allocated by each thread is not always consistent with the actual
value in global memory. In concrete terms, this causes m_rigidIndicies[]
to be corrupted such that fluid particle contacts are mapped to the
incorrect rigid bodies, causing the rigid bodies to experience collision
response forces even if there is no actual collision.