Remove soft body averageArea, fix area and Init_Volume() demo.
In practive, using (averageArea = false) in btSoftBody::updateArea()
approximately doubles the area compared to (averageArea = true).

This causes the magnitude of the volume conservation force in the
Init_Volume() demo to increase to the point that it breaks.

However, the area of a parallelogram is actually used to calculate
the per-face area. Replacing the per-face area calculation with the
area of a triangle halves the area, and also fixes the Init_Volume()
demo.

See also:
"Softby node area incorrectly calculated"
https://code.google.com/p/bullet/issues/detail?id=633

"Soft body updateConstraints() changes link rest
length, and area not being recalculated"
https://code.google.com/p/bullet/issues/detail?id=503